#2231 move to rpc.sessiontable
