+++ example/main-java/src/main/java/org/finos/vuu/person/rpc/EditPersonRecordRpcHandler.java
@@ -1,7 +1,7 @@
 package org.finos.vuu.person.rpc;
 
 import org.finos.vuu.core.table.TableContainer;
-import org.finos.vuu.net.rpc.CreateSessionTableRpcHandlerImpl;
+import org.finos.vuu.net.rpc.sessiontable.CreateSessionTableRpcHandlerImpl;
 import org.finos.vuu.net.rpc.RpcPermissionChecker;
 
 public class EditPersonRecordRpcHandler extends CreateSessionTableRpcHandlerImpl {
