+++ vuu/src/main/scala/org/finos/vuu/net/rpc/sessiontable/EndSessionRpcHandler.scala
@@ -1,6 +1,7 @@
-package org.finos.vuu.net.rpc
+package org.finos.vuu.net.rpc.sessiontable
 
 import org.finos.vuu.core.table.TableContainer
+import org.finos.vuu.net.rpc.{DefaultRpcHandler, RpcFunctionFailure, RpcFunctionResult, RpcFunctionSuccess, RpcHandler, RpcNames, RpcParams}
 
 abstract class EndSessionRpcHandlerImpl(tableContainer: TableContainer)
   extends EndSessionRpcHandler(tableContainer)
